re
re

+++ Assets/User/Yerang/Scripts/Installation/Info/FieldInfo.cs
@@ -13,7 +13,7 @@ public class FieldInfo : ScriptableObject
     public struct Stat
     {
         public float maxHp;
-        public float amountPerSec;
+        public int amountPerSec;
     }
     public Stat[] levelStat;
